+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- WordpressRelatedPostsBlock now checks to make sure productQuery is defined before rendering
+
 ## [0.1.1] - 2020-01-13
 
 ### Fixed